feat: document module, new error handler layer

+++ src/modules/story/story.controller.ts
@@ -1,10 +1,10 @@
 import { Controller, Post, Body, HttpCode, HttpStatus, Put, Param, Delete, Get } from "@nestjs/common";
 import { StoryService } from "./story.service";
-import { CreateStoryBody } from "./dtos/create-story.dto";
+import { CreateStoryBody } from "./dtos/request/create-story.dto";
 import { apiResponse } from "src/core/utils/apiResponse";
 import { CurrentUser } from "src/modules/auth/decorators/current-user.decorator";
 import { UserJwtPayload } from "src/types/UserJwtPayload";
-import { UpdateStoryBody } from "src/modules/story/dtos/update-story.dto";
+import { UpdateStoryBody } from "src/modules/story/dtos/request/update-story.dto";
 
 @Controller("story")
 export class StoryController {
@@ -24,14 +24,14 @@ export class StoryController {
     }
 
     @Put(":id")
-    async update(@Param("id") id: number, @Body() body: UpdateStoryBody) {
-        const res = await this.storyService.update({ ...body, storyId: id });
+    async update(@Param("id") id: number, @CurrentUser() user: UserJwtPayload, @Body() body: UpdateStoryBody) {
+        const res = await this.storyService.update({ ...body, storyId: id, userId: user.id });
         return apiResponse("story updated", res);
     }
 
     @Delete(":id")
-    async delete(@Param("id") id: number) {
-        const res = await this.storyService.delete({ storyId: id });
+    async delete(@Param("id") id: number, @CurrentUser() user: UserJwtPayload) {
+        const res = await this.storyService.delete({ storyId: id, userId: user.id });
         return apiResponse("story deleted", res);
     }
 }
